Compilation fix

+++ src/raknet/BitStream.h
@@ -1197,7 +1197,7 @@ namespace RakNet
 			y=0.0;
 		else
 		{
-			ReadCompressed((float)cy);
+			ReadCompressed(cy);
 			y=cy;
 			//Read(sy);
 			//y=((float)sy / 32767.5f - 1.0f);
@@ -1214,7 +1214,7 @@ namespace RakNet
 			//		return false;
 
 			//	z=((float)sz / 32767.5f - 1.0f);
-			if (!ReadCompressed((float)cz))
+			if (!ReadCompressed(cz))
 				return false;
 			z=cz;
 		}
